pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-context-mcp"
version = "1.0.0"
description = "MCP server with local RAG for Claude Code repo context"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "code-context contributors" }]
keywords = ["claude-code", "mcp", "rag", "developer-tools"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Text Processing :: Indexing",
]
dependencies = [
  "mcp>=1.0",
  "numpy>=1.26",
  "pyarrow>=14",
  "platformdirs>=4",
  "pathspec>=0.12",
  "filelock>=3.13",
  "sentence-transformers>=2.7",
  "tree-sitter>=0.22",
  "tree-sitter-language-pack>=0.7",
]

[project.optional-dependencies]
openai = ["openai>=1.30"]
watch = ["watchdog>=4"]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "watchdog>=4",
  "pyyaml>=6",
]

[project.scripts]
code-context = "code_context.cli:main"
code-context-server = "code_context.server:main"

[project.urls]
Homepage = "https://github.com/nachogeinfor-ops/code-context"
Documentation = "https://github.com/nachogeinfor-ops/code-context#readme"
Issues = "https://github.com/nachogeinfor-ops/code-context/issues"
Changelog = "https://github.com/nachogeinfor-ops/code-context/blob/main/CHANGELOG.md"
"Tool Protocol" = "https://github.com/nachogeinfor-ops/context-template/blob/main/docs/tool-protocol.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
